kernel: rename process to thread
The term "process" now refers to the abstruct concept of a resource
owner, and the term "thread" refers to the execution context and
resources related to it. Due to the change in kernel design, what
originally named "process" is now better to be called "thread".
This commit is the first step of implementing the new design.